#853 Moved the concern of test properting making from DB creator to test runner.
